Split templates.go into list and status
